+++ src/Identity.php
@@ -3,21 +3,28 @@ declare(strict_types = 1);
 
 namespace Innmind\Immutable;
 
+use Innmind\Immutable\Identity\{
+    Implementation,
+    InMemory,
+    Lazy,
+    Defer,
+};
+
 /**
  * @psalm-immutable
  * @template T
  */
 final class Identity
 {
-    /** @var T */
-    private mixed $value;
+    /** @var Implementation<T> */
+    private Implementation $implementation;
 
     /**
-     * @param T $value
+     * @param Implementation<T> $implementation
      */
-    private function __construct(mixed $value)
+    private function __construct(Implementation $implementation)
     {
-        $this->value = $value;
+        $this->implementation = $implementation;
     }
 
     /**
@@ -30,7 +37,41 @@ final class Identity
      */
     public static function of(mixed $value): self
     {
-        return new self($value);
+        return new self(new InMemory($value));
+    }
+
+    /**
+     * When using a lazy computation all transformations via map and flatMap
+     * will be applied when calling unwrap. Each call to unwrap will call again
+     * all transformations.
+     *
+     * @psalm-pure
+     * @template A
+     *
+     * @param callable(): A $value
+     *
+     * @return self<A>
+     */
+    public static function lazy(callable $value): self
+    {
+        return new self(new Lazy($value));
+    }
+
+    /**
+     * When using a deferred computation all transformations via map and flatMap
+     * will be applied when calling unwrap. The value is computed once and all
+     * calls to unwrap will return the same value.
+     *
+     * @psalm-pure
+     * @template A
+     *
+     * @param callable(): A $value
+     *
+     * @return self<A>
+     */
+    public static function defer(callable $value): self
+    {
+        return new self(new Defer($value));
     }
 
     /**
@@ -42,8 +83,7 @@ final class Identity
      */
     public function map(callable $map): self
     {
-        /** @psalm-suppress ImpureFunctionCall */
-        return new self($map($this->value));
+        return new self($this->implementation->map($map));
     }
 
     /**
@@ -55,8 +95,7 @@ final class Identity
      */
     public function flatMap(callable $map): self
     {
-        /** @psalm-suppress ImpureFunctionCall */
-        return $map($this->value);
+        return $this->implementation->flatMap($map);
     }
 
     /**
@@ -64,6 +103,6 @@ final class Identity
      */
     public function unwrap(): mixed
     {
-        return $this->value;
+        return $this->implementation->unwrap();
     }
 }